chore: Version OIDC auth provider
